Switches for modelLearner, name of param hierarchy changed
modelLearner now has switches -b and -o, which can be used to specify
that all files to be processed are background (-b), and to specify the
location to which the YAML file will be output (-o). The learner
also automatically considers directories with name "background" to
contain background images. Images in this directory, and all images
during runs with -b will not be thresholded before processing. All other
images will be thresholded in order to remove the black pixels.

Changed the param hierarchy top level name for the models to /models to
reflect the fact that the background model also sits there. Also separated
out the background model from object models.